Use package objects to apply loading workaround more reliably
- Improve doc grouping and macros
- Add more pooling ops
