+++ core/src/main/scala/torch/internal/LoadCusolver.scala
@@ -28,5 +28,4 @@ object LoadCusolver {
   } catch {
     case e: ClassNotFoundException => // ignore to avoid breaking CPU only builds
   }
-
 }
