--- core/src/main/scala/torch/nn/functional/dropout.scala
@@ -1,25 +1,0 @@
-/*
- * Copyright 2022 storch.dev
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package torch
-package nn
-package functional
-import org.bytedeco.pytorch.global.torch as torchNative
-
-def dropout[D <: DType](input: Tensor[D], p: Double = 0.5, training: Boolean = true): Tensor[D] =
-  Tensor(
-    torchNative.dropout(input.native, p, training)
-  )
