+++ core/src/main/scala/torch/nn/utils.scala
@@ -19,15 +19,16 @@ package nn
 
 import org.bytedeco.pytorch.global.torch as torchNative
 import org.bytedeco.pytorch.TensorVector
+import torch.internal.LoadCusolver
 
 object utils:
-
+  LoadCusolver
   def clipGradNorm_(
       parameters: Seq[Tensor[?]],
       max_norm: Double,
       norm_type: Double = 2.0,
       error_if_nonfinite: Boolean = false
-  ) =
+  ): Double =
     torchNative.clip_grad_norm_(
       TensorVector(parameters.map(_.native).toArray*),
       max_norm,
