+++ core/src/test/scala/torch/nn/MaxPoolSuite.scala
@@ -38,7 +38,7 @@ class MaxPoolSuite extends TensorCheckSuite {
   propertyTestUnaryOp(F.maxPool1d(_, 3), "maxPool1d", genRandTensor(shape3d))
   propertyTestUnaryOp(F.maxPool1dWithIndices(_, 3), "maxPool1dWithIndices", genRandTensor(shape3d))
 
-  inline def genRandTensor[D <: FloatNN | ComplexNN](shape: Seq[Int] = Seq(3,3)): Gen[Tensor[D]] =
+  inline def genRandTensor[D <: FloatNN | ComplexNN](shape: Seq[Int] = Seq(3, 3)): Gen[Tensor[D]] =
     Gen.oneOf(allDTypes.filter(_.isInstanceOf[D])).map { dtype =>
       torch.rand(shape, dtype = dtype.asInstanceOf[D])
     }
@@ -48,6 +48,14 @@ class MaxPoolSuite extends TensorCheckSuite {
   propertyTestUnaryOp(F.maxPool2dWithIndices(_, 3), "maxPool2dWithIndices", genRandTensor(shape4d))
 
   val shape5d = Seq(20, 16, 50, 44, 31)
-  propertyTestUnaryOp(F.maxPool3d(_, (3, 2, 2), stride=(2, 1, 2)), "maxPool3d", genRandTensor(shape4d))
-  propertyTestUnaryOp(F.maxPool3dWithIndices(_, (3, 2, 2), stride=(2, 1, 2)), "maxPool3dWithIndices", genRandTensor(shape4d))
+  propertyTestUnaryOp(
+    F.maxPool3d(_, (3, 2, 2), stride = (2, 1, 2)),
+    "maxPool3d",
+    genRandTensor(shape4d)
+  )
+  propertyTestUnaryOp(
+    F.maxPool3dWithIndices(_, (3, 2, 2), stride = (2, 1, 2)),
+    "maxPool3dWithIndices",
+    genRandTensor(shape4d)
+  )
 }
